filter_lua: extract code into flb_lua module
Extract lua related code from filter/lua plugin into a reusable module.
