+++ plugins/filter_lua/lua_config.c
@@ -47,7 +47,7 @@ struct lua_filter *lua_config_create(struct flb_filter_instance *ins,
     struct mk_list *split   = NULL;
     struct mk_list *head    = NULL;
     struct mk_list *tmp_list= NULL;
-    struct l2c_type  *l2c   = NULL;
+    struct flb_lua_l2c_type  *l2c   = NULL;
     struct flb_split_entry *sentry = NULL;
 
     /* Allocate context */
@@ -64,7 +64,7 @@ struct lua_filter *lua_config_create(struct flb_filter_instance *ins,
         return NULL;
     }
 
-    mk_list_init(&lf->l2c_types);
+    mk_list_init(&lf->l2cc.l2c_types);
     lf->ins = ins;
     lf->script = NULL;
 
@@ -120,41 +120,41 @@ struct lua_filter *lua_config_create(struct flb_filter_instance *ins,
         return NULL;
     }
 
-    lf->l2c_types_num = 0;
+    lf->l2cc.l2c_types_num = 0;
     tmp = flb_filter_get_property("type_int_key", ins);
     if (tmp) {
-        split = flb_utils_split(tmp, ' ', L2C_TYPES_NUM_MAX);
+        split = flb_utils_split(tmp, ' ', FLB_LUA_L2C_TYPES_NUM_MAX);
         mk_list_foreach_safe(head, tmp_list, split) {
-            l2c = flb_malloc(sizeof(struct l2c_type));
+            l2c = flb_malloc(sizeof(struct flb_lua_l2c_type));
 
             sentry = mk_list_entry(head, struct flb_split_entry, _head);
 
             tmp_key = flb_strndup(sentry->value, sentry->len);
             l2c->key = flb_sds_create(tmp_key);
-            l2c->type = L2C_TYPE_INT;
+            l2c->type = FLB_LUA_L2C_TYPE_INT;
             flb_free(tmp_key);
 
-            mk_list_add(&l2c->_head, &lf->l2c_types);
-            lf->l2c_types_num++;
+            mk_list_add(&l2c->_head, &lf->l2cc.l2c_types);
+            lf->l2cc.l2c_types_num++;
         }
         flb_utils_split_free(split);
     }
 
     tmp = flb_filter_get_property("type_array_key", ins);
     if (tmp) {
-        split = flb_utils_split(tmp, ' ', L2C_TYPES_NUM_MAX);
+        split = flb_utils_split(tmp, ' ', FLB_LUA_L2C_TYPES_NUM_MAX);
         mk_list_foreach_safe(head, tmp_list, split) {
-            l2c = flb_malloc(sizeof(struct l2c_type));
+            l2c = flb_malloc(sizeof(struct flb_lua_l2c_type));
 
             sentry = mk_list_entry(head, struct flb_split_entry, _head);
 
             tmp_key = flb_strndup(sentry->value, sentry->len);
             l2c->key = flb_sds_create(tmp_key);
-            l2c->type = L2C_TYPE_ARRAY;
+            l2c->type = FLB_LUA_L2C_TYPE_ARRAY;
             flb_free(tmp_key);
 
-            mk_list_add(&l2c->_head, &lf->l2c_types);
-            lf->l2c_types_num++;
+            mk_list_add(&l2c->_head, &lf->l2cc.l2c_types);
+            lf->l2cc.l2c_types_num++;
         }
         flb_utils_split_free(split);
     }
@@ -166,7 +166,7 @@ void lua_config_destroy(struct lua_filter *lf)
 {
     struct mk_list  *tmp_list = NULL;
     struct mk_list  *head     = NULL;
-    struct l2c_type *l2c      = NULL;
+    struct flb_lua_l2c_type *l2c      = NULL;
 
     if (!lf) {
         return;
@@ -179,8 +179,8 @@ void lua_config_destroy(struct lua_filter *lf)
         flb_sds_destroy(lf->buffer);
     }
 
-    mk_list_foreach_safe(head, tmp_list, &lf->l2c_types) {
-        l2c = mk_list_entry(head, struct l2c_type, _head);
+    mk_list_foreach_safe(head, tmp_list, &lf->l2cc.l2c_types) {
+        l2c = mk_list_entry(head, struct flb_lua_l2c_type, _head);
         if (l2c) {
             if (l2c->key) {
                 flb_sds_destroy(l2c->key);
