+++ plugins/filter_lua/lua_config.h
@@ -25,29 +25,17 @@
 #include <fluent-bit/flb_filter.h>
 #include <fluent-bit/flb_luajit.h>
 #include <fluent-bit/flb_sds.h>
+#include <fluent-bit/flb_lua.h>
 
 #define LUA_BUFFER_CHUNK    1024 * 8  /* 8K should be enough to get started */
-#define L2C_TYPES_NUM_MAX   16
-
-enum l2c_type_enum {
-    L2C_TYPE_INT,
-    L2C_TYPE_ARRAY
-};
-
-struct l2c_type {
-    flb_sds_t key;
-    int type;
-    struct mk_list _head;
-};
 
 struct lua_filter {
     flb_sds_t script;                 /* lua script path */
     flb_sds_t call;                   /* function name   */
     flb_sds_t buffer;                 /* json dec buffer */
-    int    l2c_types_num;             /* number of l2c_types */
     int    protected_mode;            /* exec lua function in protected mode */
     int    time_as_table;             /* timestamp as a Lua table */
-    struct mk_list l2c_types;         /* data types (lua -> C) */
+    struct flb_lua_l2c_config l2cc;   /* lua -> C config */
     struct flb_luajit *lua;           /* state context   */
     struct flb_filter_instance *ins;  /* filter instance */
 };
