+++ src/CMakeLists.txt
@@ -168,6 +168,7 @@ endif()
 if(FLB_LUAJIT)
   set(src
     ${src}
+    "flb_lua.c"
     "flb_luajit.c"
     )
 endif()
